+++ packages/solid/src/use-chat.ts
@@ -442,7 +442,11 @@ export function useChat(
             ...message,
             toolInvocations: message.toolInvocations.map(toolInvocation =>
               toolInvocation.toolCallId === toolCallId
-                ? { ...toolInvocation, result }
+                ? {
+                    ...toolInvocation,
+                    result,
+                    state: 'result' as const,
+                  }
                 : toolInvocation,
             ),
           }