+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/openai-compatible
 
+## 0.0.2
+
+### Patch Changes
+
+- fc18132: feat (ai/core): experimental output for generateText
+
 ## 0.0.1
 
 ### Patch Changes