+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/openai
 
+## 1.0.5
+
+### Patch Changes
+
+- fc18132: feat (ai/core): experimental output for generateText
+
 ## 1.0.4
 
 ### Patch Changes